requirements.txt
Django==1.11.14
django-fieldsignals==0.3.4
djangorestframework==3.8.2
MySQL-python==1.2.5
pytz==2018.5
